requirements.txt
Django==5.2.8
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.2.1